曲の情報を分割
